Refactor EvolutionaryProgressMeter module and quality-diversity.jl
This commit refactors the EvolutionaryProgressMeter module and quality-diversity.jl files.

In EvolutionaryProgressMeter, dynamic scheduling and greedy scheduling functions are added to the Evolutionary.value! method. The dynamic scheduling function calculates the value of the objective function for each element in xs using multithreading. The greedy scheduling function calculates the value of the objective function for each element in xs using multithreading and a greedy algorithm.

In quality-diversity.jl, changes are made to the initial_state, update_state!, recombine!, and mutate! functions. The initial_state function now takes an AbstractVector as the population argument. The update_state! function now takes an AbstractVector as the parents argument. The recombine! and mutate! functions now take an AbstractVector as the offspring and population arguments respectively. Additionally, the evaluate! function now takes AbstractMatrix as the objective_values argument and AbstractVector as the population argument.

These changes improve the modularity and readability of the code.